Refactor legacy unit tests to own file
